1차 리펙토링
리펙토링 테스트
