+++ Assets/Scripts/Manager/NetworkManager.cs
@@ -12,8 +12,6 @@ public class NetworkManager
     private UdpClient connectUdp = null;
     private static readonly Lazy<NetworkManager> instance = new Lazy<NetworkManager>(() => new NetworkManager());
 
-    private Queue<byte[]> tcpDataQueue = new Queue<byte[]>();
-
     public static NetworkManager Instance
     {
         get
@@ -73,8 +71,6 @@ public class NetworkManager
             Byte[] recvBuffer = new Byte[NetworkUtils.TOTAL_MAX_SIZE];
             int recvSize = clientTcpSocket.Receive(recvBuffer, NetworkUtils.TOTAL_MAX_SIZE, SocketFlags.None);
             NetworkBufferManager.Instance.AppendByTcp(recvBuffer);
-
-            EnqueueTcpData(recvBuffer);
             Debug.Log(recvBuffer);
         }
     }
@@ -159,26 +155,5 @@ public class NetworkManager
         }
     }
 
-    private void EnqueueTcpData(byte[] data)
-    {
-        lock (tcpDataQueue)
-        {
-            tcpDataQueue.Enqueue(data);
-        }
-    }
-
-    public bool TryDequeueTcpData(out byte[] data)
-    {
-        lock (tcpDataQueue)
-        {
-            if (tcpDataQueue.Count > 0)
-            {
-                data = tcpDataQueue.Dequeue();
-                return true;
-            }
-        }
-
-        data = null;
-        return false;
-    }
+    
 }
